+++ src/api/v1/auth_refresh.py
@@ -6,7 +6,7 @@ from src.core.logger import logger
 from src.services import token_refresh_service
 from src.core.configuration.config import settings
 
-router = APIRouter(tags=["Auth"])
+router = APIRouter()
 
 @router.post(
     "/refresh",
